% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_groups_col.R
\name{add_groups_col}
\alias{add_groups_col}
\title{Add group names based on experiment name}
\usage{
add_groups_col(x, samples_vector, groups_vector, samples_column,
  column_name = "groups")
}
\arguments{
\item{x}{A dataframe that contains one column with experiments names.}

\item{samples_vector}{A vector containing the experiment names.}

\item{groups_vector}{A vector containing the group names for each sample of the samples_vector.
Info for groups_vector and samples_vector can be stored in a separate dataframe and link the
corresponding columns.}

\item{samples_column}{The column name from the input dataframe that containins the sample names.
Exact match of this column will be used, so write the full name.}

\item{column_name}{The name of the new column that will include the group name. Default value is
'groups', and this column should not already exist in the dataframe used as input.}
}
\description{
Reads the experiments names from a column, and creates a new column with the group
name for each experiment. Parital matchins is supported, so groups can be matched even if you add
additional tags to the experiment names (e.g. 'log2_').
}
\examples{
 grouped_data <- add_groups(x = NS_secreted_long,
                            samples_vector = groups_df$headers_raw,
                            groups_vector = groups_df$group,
                            samples_column = "variable")

}
