% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_key.R
\name{add_key}
\alias{add_key}
\title{Add key column in the begining of your dataframe.}
\usage{
add_key(x, colname = "key")
}
\arguments{
\item{x}{Input dataframe.}

\item{colname}{Custom name for the key column. Default value is "key".}
}
\description{
Add a column with key numbers in the begining of the dataframe.
           You can use this column later as a reference to join your dataset.
}
\examples{

x <- data.frame(V1 = c("A", "B", "C"),
                V2 = c("D", "E", "F"))
add_key(x, "key_column")
}
