% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aov_padj.R
\name{calc_aov_padj}
\alias{calc_aov_padj}
\title{Perform ANOVA, calculate p-values and adjust for multiple hypothesis testing error}
\usage{
calc_aov_padj(x, sample_names, group_names, pAdj = "BH")
}
\arguments{
\item{x}{A dataframe containing the protein/gene identifier. Each row is one protein/gene and
each column is one samlple. Additional columns can be included in this dataframe. The samples
will be matched based on the \code{sample_names} vector. However because samples are identified
by partial matching (\code{grep} function) try to reduce the number of columns if you encounter an error.}

\item{sample_names}{A vector containing the sample names.}

\item{group_names}{A vector of equal length with the \code{sample_names} containing the group name
for each sample. You can store in a separate dataframe the \code{sample_names} and the \code{group_names}
 and just reference to this table.}

\item{pAdj}{Method for adjustment of p-values. See more at \code{p.adjust}. Default method is "BH".}
}
\description{
Reads a dataframe, identifies the sample groups based on the input info,
and performs ANOVA to identify rows that show difference in at least one of the groups.
P-values are adjusted for multiple hypothesis testing error, and the output is returned
as a dataframe. If the number of observations is not enought for the ANOVA test, the entry
is labeled as excluded.
}
\examples{
proc_data <- calc_aov_padj(my_data,
                           sample_names = sample_groups$samples,
                           group_names = sample_groups$groups)

}
