% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_exp_df.R
\name{calc_exp_df}
\alias{calc_exp_df}
\title{Calculate exponential values from a dataframe}
\usage{
calc_exp_df(x, base = 2, sub = "", replacement = "", exclusion = c(),
  complete_match = TRUE, append = FALSE)
}
\arguments{
\item{x}{Input data frame. It can contain also non-numeric columns.}

\item{base}{Base number for which you will calculate the exponential value.
Default value is 2.}

\item{sub}{Character string to substitute in column names. Default is empty}

\item{replacement}{Character string to replace the substituted one. Default is empty.}

\item{exclusion}{Vector containing labels from the columns that should be excluded.
Even a part of the label is enough.}

\item{complete_match}{Logical with default value \code{TRUE}. If exclusion vector
contains exactly the column names of the excluded columns, leave the
\code{complete_match} to \code{TRUE}. For partial matching of the columns switch
to \code{FALSE}.}

\item{append}{Logical with default value \code{FALSE}. If set \code{TRUE} processed
data will be added in the starting dataframe. If \code{TRUE}, the processed columned
will be merged only with the excluded ones.}
}
\description{
Reads a data frame and calculates the exponential value (basex) for
the defined columns. Non-numeric columns can be excluded by adding the column name
in exclusion vector. For partial matching of column names use
\code{complete_match = FALSE}. See \code{calc_exp_vec}.
}
\examples{
y1 <- calc_exp_df(x = my_data,
                  base = 2,
                  sub = "norm_log_",
                  replacement = "norm_",
                  exclusion = colnames(y)[1:3],
                  complete_match = TRUE)

}
