% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_exp_vec.R
\name{calc_exp_vec}
\alias{calc_exp_vec}
\title{Calculate exponential base^x from a vector}
\usage{
calc_exp_vec(x, base = 2)
}
\arguments{
\item{x}{Numeric vector.}

\item{base}{Numeric element that you will raise \code{x}.}
}
\description{
Calculate exponential basex. Reads a numeric
element convert to the power specified.
}
\examples{
x <- 4
base <- 2
calc_exp(a, x)

v <- seq(1:5)
calc_exp(v)
}
