% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fold_change_df.R
\name{calc_fold_change_df}
\alias{calc_fold_change_df}
\title{Calculate fold change}
\usage{
calc_fold_change_df(x, mean_A, mean_B, colname = "log2_fold_difference")
}
\arguments{
\item{x}{Data frame containing input values.}

\item{mean_A}{Character string describing the column name of the mean for the
group A. Partial matching is acceptable.}

\item{mean_B}{Character string describing the column name of the mean for the
group B. Partial matching is acceptable.}

\item{colname}{Single character vector, with the label of the column with the calculated values.
Default value is "log2_fold_difference".}
}
\description{
Reads a data frame, identifies the columns containing group mean values, and
calculates the log2 fold difference. Output is saved in a new column of the same data frame
labeled as “log2_fold_change_AB”. If means are not calculated, used \code{calc_means_df} function first.
}
\examples{
proc_data <- calc_fold_change_df(data, "TIC_Pat", "TIC_Ctr")
}
