% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kruskal_test.R
\name{calc_kruskal_test}
\alias{calc_kruskal_test}
\title{Perform Kruskal-Wallis rank sum test in multiple rows and correct p-values}
\usage{
calc_kruskal_test(x, groupA, groupB, complete_match = FALSE,
  padj_method = "BH")
}
\arguments{
\item{x}{Input dataframe}

\item{groupA}{Character vector containing identifiers from samples of groupA. You can also use partial matching.}

\item{groupB}{Character vector containing identifiers from samples of groupB. You can use partial matching.}

\item{complete_match}{Logical with default value FALSE, allowing partial matching for groupA and groupB.
If you want to use exact match, switch to TRUE.}

\item{padj_method}{Method for adjustment of p-values for multiple hypothesis testing error (see
\code{link{p.adjust}} for more info). Default method is "BH".}
}
\description{
Perform Kruskal-Wallis rank sum test in multiple rows and correct p-values.
}
