% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log2_df.R
\name{calc_log2_df}
\alias{calc_log2_df}
\title{Calculate log2 values in a dataframe}
\usage{
calc_log2_df(x, log_lab, rm_zero = TRUE, exclusion = c(),
  complete_match = TRUE, append = FALSE)
}
\arguments{
\item{x}{Dataframe containing only numeric values.}

\item{log_lab}{Character tag to be added in front of every column name of the dataframe (e.g. "Log2_".}

\item{rm_zero}{If \code{TRUE}, missing values will remain \code{NA}. If set to \code{FALSE}
missing values will be replaced by 0. By default \code{rm_zero = TRUE}.}

\item{exclusion}{Vector containing labels from the columns that should be excluded. Even a part of the label is enough.}

\item{complete_match}{Logical with default value \code{TRUE}. If exclusion vector contains exactly
the column names of the excluded columns, leave the \code{complete_match} to \code{TRUE}. For partial
matching of the columns switch to \code{FALSE}}

\item{append}{Logical with default value \code{FALSE}. If set \code{TRUE} processed data will be
added in the starting dataframe. If \code{FALSE}, the processed columned will be merged only with
the excluded ones.}
}
\description{
Reads a data-frame and calculates log2 for each value. Non-numeric columns that
should be transformed can be excluded by adding the column name in exclusion vector. For partial
matching of column names use \code{complete_match = FALSE}. Transformed columns can be renamed
using the \code{log_lab} argument. Missing values can be either replaced with 0 or removed by
adjusting the \code{rm_zero} argument.
}
\examples{
calc_log2_df(x, log_lab = "my_label", rm_zero = FALSE)
}
