% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log2_vec.R
\name{calc_log2_vec}
\alias{calc_log2_vec}
\title{Calculate log2 from a vector}
\usage{
calc_log2_vec(x)
}
\arguments{
\item{x}{Numeric element or vector for which the log2 value will be calculated.}
}
\description{
Reads a numeric element or a vector and returns its log2 value. Output is numeric.
Missing values will be replaced with 0. You can use \code{convert_to_vec} or \code{convert_to_df}
to convert 0 back to NA.
}
\examples{
x <- seq(from = 100, to = 1000, by = 100)
x
calc_log2_vec(x)
}
