% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_TIC.R
\name{calc_TIC}
\alias{calc_TIC}
\title{Calculate sum of peptide intensities}
\usage{
calc_TIC(x, col_id, protein_lab, custom_lab = "TIC_")
}
\arguments{
\item{x}{Dataframe containing list of protein identifiers and intensity values for individual peptides.
Each row contains peptide intensities.}

\item{col_id}{Character string used to identify the columns containing peptide intensity.}

\item{protein_lab}{Column name of the column containing the protein identifiers.}

\item{custom_lab}{Custom character string to be added in front of every column name.
Default value it "TIC_".}
}
\description{
Reads a table containing a list of petides intensities and sums the peptide intensities
for each protein. Column containing protein identifiers should be included in the dataset.
}
\examples{
TIC <- calc_TIC(x, col_id = "normalized_int", protein_lab = "Protein")
}
