% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ttest_padj.R
\name{calc_ttest_padj}
\alias{calc_ttest_padj}
\title{Perform t-test and adjust p-values}
\usage{
calc_ttest_padj(x, id_groupA = "_groupA", id_groupB = "_groupB",
  min_A = 3, min_B = 3, t.test_var.equal = FALSE, pAdj_method = "BH",
  output_name = "processed p-values.txt")
}
\arguments{
\item{x}{Dataframe containing quantitative values.}

\item{id_groupA}{Identifier for the samples of the first study group.}

\item{id_groupB}{Identifier for the samples of the second study group.}

\item{min_A}{Minimum number of samples to be tested from group A.}

\item{min_B}{Minimum number of samples to be tested from group B.}

\item{t.test_var.equal}{Assumption for equal variance in t-test. Default is \code{FALSE}.
Set \code{TRUE} is you assume equal variance in the set.}

\item{pAdj_method}{Method for multiple hypothesis testing of the p-values. Default is "Benjamini-Hochberg"
method "BH". Run \code{?p.adjust} for more info.}

\item{output_name}{Name of the output file containing the results. This output will be saved in your
working directory. For this set the desired working directory before running the function. Also
the output will be printed, so you can save it in a variable.}
}
\description{
Reads a dataframe, recognized protein groups, calculates protein identification frequency,
performes t-test and calculates p-value, and adjusts p-value for multiple hypothesis testing error.
The starting dataframe can also contain columns with text. Experimental groups are identified based
on specific character string for each group. If missing values are converted to \code{NA}, they will
be omitted from the t-test. Minimum number of samples to be used in t-test can be defined by the user.
Default value is 3 samples per group. Two tailed t-test is used without assuming equal variance.
Equal variance can be selected by user using \code{t.test_var.equal} argument. P-values are
adjusted using "Benjamini-Hochberg" method "BH". To use a different method, change \code{pAdj_method}
parameter (see \code{?p.adjust} for more info. Results are saved in your working directory and can
be assigned to a variable for further processing.
}
\examples{
calc_ttest_padj(norm_data, id_groupA = "log2_disease", id_groupB = "log2_controls")
}
