% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_df.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Convert selected values from a dataframe}
\usage{
convert_to_df(x, lab = "", value = 0, to = NA, exclusion = c(),
  complete_match = TRUE)
}
\arguments{
\item{x}{A dataframe with values.}

\item{lab}{Label to be added in each column after conversion. By default nothing is added.}

\item{value}{Value to be changed. Default is 0.}

\item{to}{Target value. Default is \code{NA}.}

\item{exclusion}{Vector containing labels from the columns that should be excluded. Even a partial match is enough.}

\item{complete_match}{Logical with default value \code{TRUE}. If exclusion vector contains exactly
the column names of the excluded columns, leave the \code{complete_match} to \code{TRUE}.
For partial matching of the columns switch to \code{FALSE}.}
}
\description{
Convert selected values to target value. Input and output are both dataframes.
By default 0 is replaced with \code{NA}. Non-numeric column names should be included in the exclusion
argument. Columns can be selected either by exact name matching or by partial matching. For
partial matching turn \code{complete_match = FALSE}.
}
\examples{
new_data <- convert_to_df(x, value = 0, to = NA, exclusion = c(“Protein”, “pept”))
}
