% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_indexes.R
\name{get_col_indexes}
\alias{get_col_indexes}
\title{Get column indexes of a dataframe}
\usage{
get_col_indexes(x, id)
}
\arguments{
\item{x}{A dataframe to with the columns of interest}

\item{id}{Vector contaning character identifiers for the columns of interest. You can match
the column names partially (see example).}
}
\description{
Reads a dataframe and returns the column indexes that match to a specific pattern.
If input is not a dataframe, returns an error. You can set the first argument as.data.frame(x).
}
\examples{
x <- data.frame(one = 1, two = 2, three = 3, four = 4)
vect <- c("o", "wo")

get_col_indexes(x, "wo")
get_col_indexes(x, vect)

}
