% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDX_ltseries_uptake_ratio.R
\name{HDX_ltseries_uptake_ratio}
\alias{HDX_ltseries_uptake_ratio}
\title{Calculate deuterioum uptake ration between the multiple time points over the last.}
\usage{
HDX_ltseries_uptake_ratio(x, protein_col, peptide_col = "Sequence",
  uptake_col = "Uptake", time_col = "Exposure")
}
\arguments{
\item{x}{A dataframe with the input values.}

\item{protein_col}{Name of the column containing the multiple proteins compared.}

\item{peptide_col}{Name of the column containing the peptide sequence.}

\item{uptake_col}{Name of the column containing the deuterium uptake values.}

\item{time_col}{Name of the column containing the timepoint of the measurement.}
}
\description{
Calculate deuterioum uptake ration between the multiple time points over the last,
by dividing the uptake value from each timepoint with the last time point of the series.
This division will be performed for each peptide of each protein.
}
\examples{
x <- data
protein_col <- "mutant"
peptide_col <- "Sequence"
uptake_col <- "Uptake"
time_col <- "Exposure"

time_series_ratios <- HDX_ltseries_uptake_ratio(x = data, protein_col = "mutant")

}
