% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwd_enrichment.R
\name{kwd_enrichment}
\alias{kwd_enrichment}
\title{Test for keyword enrichment in the differentially abundant/expressed ids}
\usage{
kwd_enrichment(kwd_col_test, kwd_col_bg, kwd_vector, kwd_sep = "; ",
  padj_method = "bonferroni")
}
\arguments{
\item{kwd_col_test}{A vector containing keywords for the differentially abundant/expressed ids. Multiple
keywords can be included for each cell, separated by and identifier (by default '; '). Each
element of the vector can contain keywords for a specific protein. You can add filter your
dataset to the ids of interest, add annotation using \code{append_cols_df()} function, and
add assign the corresponding column (e.g. \code{kwd_col_test = my_data$keywords}).}

\item{kwd_col_bg}{Similar with the \code{kwd_col_test}, is a vector containing keywords for the ids of
background genome/proteome.}

\item{kwd_vector}{A character vector containing the keywords that you want to test for significance.
Each keyword will be matched exactly, but there is an elasticity for the use of upper
or lower case characters.}

\item{kwd_sep}{In case that multiple keywords are provided for each protein/gene, you can choose the
keyword seperator. Default value is "; ".}

\item{padj_method}{Keywords will be tested for significance using Fisher's exact test. P-values will
be further adjusted, by default using "bonverroni" method. See \?p.adjust() for more info.}
}
\description{
Reads a list of keywords and tests for enrichment in a subset of genes/proteins over
the background. As background can be selected the total list of ids or the complete genome/proteome.
Significant enrichment is tested using Fisher's exact test. Calculated p-values are furhter adjusted
for multiple hypothesis testing error. See arguments for more info.
}
\examples{
kwd_enrichment(kwd_col_test = significant_ids$keywords,
               kwd_col_bg = background_set$keywords,
               kwd_vector = c("proteolysis", "transport", "immune system"))

}
