% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwd_enrichment_multi.R
\name{kwd_enrichment_multi}
\alias{kwd_enrichment_multi}
\title{Keyword enrichment for multiple categories}
\usage{
kwd_enrichment_multi(x, categories, test_keywords, grouping_col,
  keywords_col = "keywords", padj_method = "BH")
}
\arguments{
\item{x}{A dataframe with the values.}

\item{categories}{A character vector containing the categories that will be tested.}

\item{test_keywords}{A caracter vector with the keywords to be tested.}

\item{grouping_col}{A character string matching exactly the column of the dataframe \code{x}
that contains the labaled for each category.}

\item{keywords_col}{A character string matching exactly the column of the dataframe \code{x}
that contains the keywords.}

\item{padj_method}{Method to adjust the p-values of the enrichment test (Fisher exact test).
See \code{link{p.adjust}} for more info.}
}
\description{
Reads multiple category identifiers and performes a keyword enrichment
   for each combination of test/background group (see \code{\link[komics]{kwd_enrichment}}).
   The output is a list of tables.
}
\examples{
\donttest{
  test_keywords <- c("protease",
                     "solute binding protein",
                     "oxidoreductase",
                     "transport",
                     "virulence")

  categories <- c("stable", "decreasing", "increasing")
  category_comparison <- kwd_enrichment_multi(x = merged_data,
                                              categories = categories,
                                              grouping_col = "Grouping_col",
                                              test_keywords = test_keywords)

}
}
