% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_columns.R
\name{move_columns}
\alias{move_columns}
\title{Move columns in a dataframe}
\usage{
move_columns(x, columns, position)
}
\arguments{
\item{x}{A dataframe.}

\item{columns}{A vector containing column names. You can use partial matching and regular expressions.}
}
\description{
Re-arrange a dataframe by moving columns. You can use partial match to
             select the columns of interest (See \code{\link{get_col_indexes}}).
}
\examples{
x <- data.frame(V1 = c("a", "b", "c"),
                V2 = c("A", "B", "C"),
                Z3 = c("c", "d", "e"))

 move_columns(x, "V1", 2)
 move_columns(x, "1$", 2)
 move_columns(x, "^Z", "start")

}
