% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_to_max.R
\name{norm_to_max}
\alias{norm_to_max}
\title{Normalize rows to the maximum value}
\usage{
norm_to_max(x, columns = "..all..", label = "nmax_",
  complete_match = FALSE, sub_min = TRUE, na.rm = TRUE,
  drop_input = FALSE)
}
\arguments{
\item{x}{A dataframe with the values to be normalized.}

\item{columns}{Vector containing column names to be standardized. By default all columns will be selected
\code{"..all.."}. You can add a vector with the column names, or a part of the column name that is unique
and use partial matching to identify columns (e.g. "values_"). If complete matching is used, specify
\code{complete_match} argument to TRUE.}

\item{label}{Label tag to be added in each collumn. Default value is "nmax_".}

\item{complete_match}{Logical with default value \code{FALSE} that allows partial matching of the column
names for selection of the column for normalizaiton. If set to \code{TRUE}, columns that match exactly with
the input will be used.}

\item{sub_min}{Logical with default value \code{TRUE}. By default the calculation is
\code{(x - x_min) / (x_max - x_min)}. If \code{sub_min} is set to \code{FALSE} the normalization will be
performed as: \code{x / x_max}}

\item{na.rm}{Action for missing values when calculating min and max values. Defalut is \code{TRUE}}

\item{drop_input}{Select if you want to append the normalized data into the input dataframe or keep only them.
Logical with default value \code{FALSE}, meaning that the processed values will be added to the input.}
}
\description{
Reads rows and normalized the values based on the maximum of the row (range 0-1).
}
\examples{
data <- data.frame(L1 = c("a", "b", "c", "d", "e", "f", "g", "h"),
                   V1 = c(1, 2, 3, 6, 1, 3, 4, 4),
                   V2 = c(3, 4, 5, 3, 5, 5, 3, 2),
                   V3 = c(6, 7, 4, 2, 3, 2, 7, 6))
 norm_to_max(data, columns = "V")
 norm_to_max(data, columns = "V", sub_min = FALSE)

}
