% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kwd_bars.R
\name{plot_kwd_bars}
\alias{plot_kwd_bars}
\title{Plot keywords after enrichment test as bargraph}
\usage{
plot_kwd_bars(x, group_vec, save_output = FALSE, x_discrete = FALSE)
}
\arguments{
\item{x}{A list containing the keyword enrichment output from \code{link[komics]{kwd_enrichment_multi}}
function. You remove elements of this list, as long as you don't change the column names.}

\item{group_vec}{A vector containing the labels of each group. Should have the same length
with the \code{x} argument.}

\item{save_output}{Defalut value is FALSE. If a character string is added, the output will be saved
in your working directory with the specified name, as tab delimited file.}

\item{x_discrete}{Default valus is FALSE. Use the \code{kwd_vector } from \code{\link[komics]{kwd_enrichment_multi}}
to order the \code{x} axis, accordingly.}
}
\description{
Reads the output of \code{link[komics]{kwd_enrichment_multi}}, along with
        a vector with the categories names and plots a linegraph of the keyword ratio between
        test group and the background.
        You don't have any options to modify the appearence of the graph, in this version.
}
\examples{
\donttest{

plot_kwd_bars(x = kwd_enrichment_multi_output,
               group_vec = categories_vector)
}
}
