% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kwd_lines.R
\name{plot_kwd_lines}
\alias{plot_kwd_lines}
\title{Plot keywords after enrichment test as lines}
\usage{
plot_kwd_lines(x, group_vec)
}
\arguments{
\item{x}{A list containing the keyword enrichment output from \code{link[komics]{kwd_enrichment_multi}}
function. You remove elements of this list, as long as you don't change the column names.}

\item{group_vec}{A vector containing the labels of each group. Should have the same length
with the \code{x} argument.}
}
\description{
Reads the output of \code{link[komics]{kwd_enrichment_multi}}, along with
        a vector with the categories names and plots a linegraph of the keyword ratio between
        test group and the background.
        You don't have any options to modify the appearence of the graph, in this version.
}
\examples{
\donttest{

plot_kwd_lines(x = kwd_enrichment_multi_output,
               group_vec = categories_vector)
}
}
