% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CV_vec.R
\name{calc_CV_vec}
\alias{calc_CV_vec}
\title{Calculate CV for one vector}
\usage{
calc_CV_vec(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Calculate CV for one vector
}
\examples{
x <- seq(1:10)
calc_CV_vec(x)
}
