% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_means_df.R
\name{calc_means_df}
\alias{calc_means_df}
\title{Calculate means}
\usage{
calc_means_df(x, groups, na.rm = TRUE)
}
\arguments{
\item{x}{Data frame with the input values.}

\item{groups}{List containing column identifiers for the different groups. The name of each
element will be used as column name containing the mean. Each element is a vector containing
possible identifiers for the columns. Identifiers can match partially with the columns of interest.}

\item{na.rm}{Logical with default value TRUE, indicating whether the NA values should be taken into account
when calculating the means. See \code{\link{mean}} function.}
}
\description{
Calculate mean values for the specified columns of a dataframe. Output is appended
into the initial dataframe.
}
\examples{
groups <- list(mean_A = "group A",
               mean_B = c("^B", “groupB”)
data <- calc_means_df(x = input, groups = groups)
}
